ipavault: simplify ipalib import process
Drop the use of a function to maybe initialize the API based on whether
it's been initialized before, tracked with a global variable.

Instead import ipalib in the normal way. This screw up logging if ipalib
is imported before Flask has had a chance to initialize logging. We'll
fix that in a subsequent commit.