Explicitly initialize logging when hosted by the Flask development server
Flask, werkzeug and ipalib all initialize their logging if nothing else
has initialized logging. But ipalib initializes logging at import time,
and since we import it before we create our Flask app, ipalib wins.

We could use an implementation detail of ipalib to cause it to not
initialize logging. But in principle any other module could initialize
logging before Flask does so.

So instead we explicitly initailize logging, if we detect we're hosted
by the Flask development server, as early as possible.

When running elsewhere, we configure logging a bit later. That process
hooks into gunicorn's error log if we detect it exists; that way,
logging can be configured via gunicorn's logging config.