log_config: use basicConfig to configure logging
I admit defeat. Too many things in Python make the assumption that if
the root logger has no handlers, then it's OK for them to go off and
configure their own logging. So we need to configure the root handler
after all.

We maintain the courtesy of trying to figure out our log level based on
the environment in which we're hosted; that is still a useful
convenience, it's just that now we unconditionally log to stderr rather
than try to output to a logger owned by our host.

The only downside of this approach is that when running in development,
messages from the werkzeug logger (e,g, request logs) are prefixed by
%(level)s:%(name)s... oh well.