Get clock parsing tests passing with TZ=UTC
The tests failed from within the GitHub actions runner because the test
data I calculated is wrong: the expected POSIX timestamp I provided was
in local time (Europe/London) and the timestamp being parsed just so
happened to be during British Summer Time!

The tests still fail with TZ=Europe/London, they will be fixed in a
subsequent commit.