test_client: use the get method when asserting request properties
The mapping type that werkzeug uses for headers, cookes, form data, etc.
throws a (derived class of) KeyError when the requested item is not
present. The message for this exception is designed to bubble up and
result in a 400 'server could not understand request' error, which makes
sense when it's thrown from an HTTP request handler function. But in
tests it's just a weird and confusing error message.

Using the get method of the mapping is more appropriate for test
assertions.