test_client: stop using ordered requests incorrectly
I misunderstood how to use them. The problem is that you're not supposed
to call httpserver.check() at the end of your test; you're supposed to
call it after each request. And each time you call it, it only checks
assertions based on the sequence of calls the httpserver has already
seen.

This means that calling it at the end of a test _won't_ fail the test if
the 'when' test stopped earlier. Specifically, if we modify
test_login_logout and comment out the call to client.logout(), the test
will still pass!

Since we're not relying on httpserver.check(), we can use the unordered
expect_request method to configure httpserver. Instread we'll set a flag
and check it to ensure that there was a request to the logout URL.