Run container tests in CI
We'd need poetry to install dependencies in the container_image job so
that we can run the container tests. So there's no advantage in having
two jobs any more.

Therefore I've combined them back into a single job. This has the
advantage of removing the discrepancy where a local image build would
build the wheel inside the container, but in CI we'd build outside; now
we go back to always building inside.